refactor: factorize loop utils for UI
